Permissions to create/remove categories

+++ final/actions/admin/remove_category.php
@@ -3,6 +3,11 @@
 include_once ("../../config/init.php");
 include_once ($BASE_DIR."database/categories.php");
 
+//Check permissions
+$member = getMemberByUsername($_SESSION['username']);
+if($member['privilege_level'] != "Administrator")
+    die('You don\'t have permissions for perform this action...');
+
 $id = $_POST['id'];
 $id = deleteCategory($id);
